refactor(core): move types into @flareapp/core
