+++ packages/core/src/index.ts
@@ -1,2 +1,14 @@
-// eslint-disable-next-line unicorn/require-module-specifiers
-export {};
+export type {
+    AttributeValue,
+    Attributes,
+    Config,
+    EntryPointHandler,
+    Framework,
+    Glow,
+    MessageLevel,
+    OverriddenGrouping,
+    Report,
+    SdkInfo,
+    SpanEvent,
+    StackFrame,
+} from './types';
