+++ packages/js/package.json
@@ -46,6 +46,7 @@
         "release": "release-it"
     },
     "dependencies": {
+        "@flareapp/core": "0.1.0",
         "error-stack-parser": "^2.0.2"
     },
     "devDependencies": {
